setup: allow disabling all htaccess fiddling
Distribution-packaged ownCloud, and other deployments where
the admin/deployer has sufficient control over the system, can
do this stuff much better. For instance, a sensible packaged
OC will use a directory that is not accessible via the web
server *at all* as the data directory (rendering it unnecessary
to attempt to 'protect' it with a .htaccess file) and will
include the 403/404 ErrorDocument directives, along with other
appropriate Apache configuration directives, in a config file
stored (again) outside the web-served locations.

This makes both of setup's attempts to fiddle about with
.htaccess stuff optional - they won't happen if the config
directive check_for_working_htaccess is set to false. We
could maybe keep that directive with its current meaning and
add a new one for 'don't mess with htaccess files at all',
but I think anyone who's setting it to false probably wants
this behaviour.